Add some remote (multi-player) works (Chat, User::MakeEnterFieldPacket ... etc.). Fix that broadcasting OutPacket causes error code 38 (replicating a new buffer). Rename some constants to enums. Fix codes related to avatar. Move some stats from GA_Character to GW_CharacterStat. Add "#mob" command for testing. Fix some bugs of Life Controller (UpdateHeap correctly). Fix SocketBase that close socket improperly (send packet even after the socket has been closed). Change std::mutex in User to std::recursive_mutex (since MakeEnterFieldPacket call a function belonged to QWUQuestRecord that lock user also).

+++ WvsGame/SkillInfo.cpp
@@ -129,7 +129,7 @@ void SkillInfo::IterateSkillInfo()
 			continue;
 		aRoot.push_back(node);
 	}
-	m_nRootCount = aRoot.size();
+	m_nRootCount = (int)aRoot.size();
 	for (auto& node : aRoot)
 	{
 		nRootID = atoi(node.Name().c_str());
